refactor : 컨트롤러에서 Input 받는 뷰를 클래스 분리
